Continue break-lock feature
