+++ setup.cfg
@@ -71,10 +71,10 @@ omit = tests/*
 relative_files = true
 
 [flake8]
-ignore = W503
+ignore =
 max-line-length = 120
 exclude =
-    build,dist,.git,.idea,.cache,.tox,.eggs,venv
+    build,dist,.git,.idea,.cache,.tox,.eggs,
     ./src/vorta/__init__.py
 
 [tox:tox]
@@ -96,3 +96,13 @@ commands=flake8 src tests
 
 [pycodestyle]
 max_line_length = 120
+
+[pylint.master]
+extension-pkg-whitelist=PyQt5
+load-plugins=
+
+[pylint.messages control]
+disable= W0503,W0511,C0301,R0903,R0201,W0212,C0114,C0115,C0116,C0103,E0611,E1120,C0415,R0914,R0912,R0915
+
+[pylint.format]
+max-line-length=120
