+++ src/vorta/application.py
@@ -9,6 +9,7 @@ from PyQt5.QtWidgets import QMessageBox
 from vorta.borg.borg_thread import BorgThread
 from vorta.borg.create import BorgCreateThread
 from vorta.borg.version import BorgVersionThread
+from vorta.borg.break_lock import BorgBreakThread
 from vorta.config import TEMP_DIR
 from vorta.i18n import init_translations, translate
 from vorta.models import BackupProfileModel, SettingsModel, cleanup_db
@@ -73,6 +74,7 @@ class VortaApp(QtSingleApplication):
         self.backup_finished_event.connect(self.backup_finished_event_response)
         self.backup_cancelled_event.connect(self.backup_cancelled_event_response)
         self.message_received_event.connect(self.message_received_event_response)
+        self.backup_log_event.connect(self.react_to_log)
         self.aboutToQuit.connect(cleanup_db)
         self.set_borg_details_action()
         self.installEventFilter(self)
@@ -171,3 +173,38 @@ class VortaApp(QtSingleApplication):
         msg.setInformativeText(self.tr("Vorta was unable to locate a usable Borg Backup binary."))
         msg.setStandardButtons(QMessageBox.Ok)
         msg.exec()
+
+    def react_to_log(self, mgs, context):
+        """
+        Trigger Vorta actions based on Borg logs. E.g. repo lock.
+        """
+        msgid = context.get('msgid')
+        if msgid == 'LockTimeout':
+            profile = BackupProfileModel.get(name=context['profile_name'])
+            repo_url = context.get('repo_url')
+            msg = QMessageBox()
+            msg.setStandardButtons(QMessageBox.Yes | QMessageBox.No)
+            msg.setText(
+                self.tr(
+                    f"The repository at {repo_url} might be in use by another computer. Override it and continue?"))
+            msg.accepted.connect(lambda: self.break_lock(profile))
+            msg.setWindowTitle(self.tr("Repository In Use"))
+            self._msg = msg
+            msg.show()
+        elif msgid == 'LockFailed':
+            repo_url = context.get('repo_url')
+            msg = QMessageBox()
+            msg.setText(
+                self.tr(
+                    f"You do not have permission to access the repository at {repo_url}. Gain access and try again."))  # noqa: E501
+            msg.setWindowTitle(self.tr("No Repository Permissions"))
+            self._msg = msg
+            msg.show()
+
+    def break_lock(self, profile):
+        params = BorgBreakThread.prepare(profile)
+        if not params['ok']:
+            self.set_progress(params['message'])
+            return
+        thread = BorgBreakThread(params['cmd'], params, parent=self)
+        thread.start()
