+++ src/vorta/borg/borg_thread.py
@@ -229,20 +229,20 @@ class BorgThread(QtCore.QThread, BackupProfileMixin):
                 for line in stderr.split('\n'):
                     try:
                         parsed = json.loads(line)
-                        context = {
-                            'msgid': parsed.get('msgid'),
-                            'repo_url': self.params['repo_url'],
-                            'profile_name': self.params.get('profile_name'),
-                            'cmd': self.params['cmd'][1]
-                        }
+
                         if parsed['type'] == 'log_message':
+                            context = {
+                                'msgid': parsed.get('msgid'),
+                                'repo_url': self.params['repo_url'],
+                                'profile_name': self.params.get('profile_name'),
+                                'cmd': self.params['cmd'][1]
+                            }
                             self.app.backup_log_event.emit(
                                 f'{parsed["levelname"]}: {parsed["message"]}', context)
                             level_int = getattr(logging, parsed["levelname"])
                             logger.log(level_int, parsed["message"])
                         elif parsed['type'] == 'file_status':
-                            self.app.backup_log_event.emit(
-                                f'{parsed["path"]} ({parsed["status"]})', context)
+                            self.app.backup_log_event.emit(f'{parsed["path"]} ({parsed["status"]})', {})
                         elif parsed['type'] == 'archive_progress':
                             msg = (
                                 f"{self.category_label['files']}: {parsed['nfiles']}, "
